pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smart-qr-tools"
version = "0.1.0"
description = "Command-line utility to create text and UPI QR codes"
authors = [{ name = "Your Name", email = "you@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["qr", "upi", "cli", "python"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "qrcode[pil]>=7.4.2",
    "segno>=1.5.0",          # SVG & micro-QR
    "pillow>=10.0.0",
]

[project.scripts]
smartqr = "smartqr.cli:main"
